Renamed stuff to make things clearer
